requirement.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
django-environ==0.9.0
frozenlist==1.3.3
idna==3.4
instaloader==4.9.5
magic-filter==1.0.9
multidict==6.0.4
pytz==2022.7.1
requests==2.28.2
urllib3==1.26.14
yarl==1.8.2
